# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Shared Key Value List System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/head/{id}": {
            "get": {
                "description": "test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a head by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Head ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.List"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.List": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nextPageID": {
                    "type": "integer"
                },
                "pageCount": {
                    "type": "integer"
                }
            }
        },
...
    }
}